Add SimpleBuildCacheKey to the library
